+++ tests/tools/test_function.py
@@ -17,8 +17,8 @@ from substratools.task_resources import StaticInputIdentifiers
 from substratools.task_resources import TaskResources
 from substratools.workspace import FunctionWorkspace
 from tests.tools import utils
-from tests.utils import InputIdentifiers
-from tests.utils import OutputIdentifiers
+from tests.tools.utils import InputIdentifiers
+from tests.tools.utils import OutputIdentifiers
 
 
 @pytest.fixture(autouse=True)
