+++ tests/tools/test_metrics.py
@@ -14,8 +14,8 @@ from substratools import save_performance
 from substratools.task_resources import TaskResources
 from substratools.workspace import FunctionWorkspace
 from tests.tools import utils
-from tests.utils import InputIdentifiers
-from tests.utils import OutputIdentifiers
+from tests.tools.utils import InputIdentifiers
+from tests.tools.utils import OutputIdentifiers
 
 
 @pytest.fixture()
