+++ tests/tools/test_workflow.py
@@ -11,8 +11,8 @@ from substratools.task_resources import TaskResources
 from substratools.utils import import_module
 from substratools.workspace import FunctionWorkspace
 from tests.tools import utils
-from tests.utils import InputIdentifiers
-from tests.utils import OutputIdentifiers
+from tests.tools.utils import InputIdentifiers
+from tests.tools.utils import OutputIdentifiers
 
 
 @pytest.fixture
